core: GetImageInfo to use GetVolumeInfo
GetImageInfoVDSCommand is an IRS command that executes getVolumeInfo
on the SPM.
Additionally, we have the GetVolumeInfoVDSCommand that executes getVolumeInfo on
the passed host in the parameters.

Currently GetVolumeInfoVDSCommand doesn't return the actual image info, but only if it
managed to get the volume information. As we need to use it to get the actual info
instead of duplicating the code or exporting it to a helper class we can change GetImageInfo
to use GetVolumeInfo and to pass it the SPM host id, maintaining the same behavior with less
code.

Change-Id: Iebfd4678e2422352567cba61ca363348867541c2
Signed-off-by: Liron Aravot <laravot@redhat.com>