core: rm DoesImageExistVDSCommand
DoesImageExistVDSCommand is a subclass of GetImageInfoVDSCommand
and is used to determine if the image exists.
The difference between this command and GetImageInfo is that GetImageInfo
parses to return value into a object.
As this command is executed as part of the validation when doing import,
we can simply use GetImageInfo and delete DoesImageExistVDSCommand.

Change-Id: I9b431ee842cf87b8b6378ef2641d64d2fa68df42
Signed-off-by: Liron Aravot <laravot@redhat.com>