webadmin: Look for sound device by general type
The sound device in the managed device map of a VM must be located by
VmDeviceGeneralType.SOUND, not by VmDeviceType.SOUND.

The code that was fixed in
InstanceTypeManager.postDoUpdateManagedFieldsFrom(VmBase) is called when
user edits a VM and NEXT_RUN snapshot exists. After getting the
information about NEXT_RUN snapshot, UI controls in the dialog are
updated accordingly. The "Sound card" checkbox is always unchecked,
because the code looks for a sound device incorrectly and cannot find
it. When user clicks "OK" in the dialog, the difference between actual
sound device configuration and the checkbox state is detected and
confirmation dialog is displayed.

Change-Id: I179bb4623815a90b5468e25e3432351e5458e04a
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1356568
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>