Packaging: spec: Make engine-setup require jre 1.8
oVirt engine 3.6 could have worked with either jre 1.7 or 1.8. 4.0
requires 1.8. But engine-setup checks if required java version is
available prior to upgrading the engine. So engine-setup will fail
when upgrading from 3.6 with 1.7. Fix this by requiring the correct
version by the package setup-plugin-ovirt-engine-common, which is
where the check is done.

Change-Id: I059daa2533e62ba3370608a507797f2b898f3235
Bug-Url: https://bugzilla.redhat.com/1350397
Signed-off-by: Yedidyah Bar David <didi@redhat.com>