webadmin: improve host kernel configuration message
In order to modify kernel boot parameters from webadmin (hosts tab ->
edit host -> kernel tab), the admin must reinstall (via webadmin) and
then manually reboot the host after modifying the kernel cmdline.

Improving the message appears in the kernel boot parameters dialog
so that it will be clear that this procedure is needed in order to
make the changes to take effect.

Change-Id: I619eb9ac697db8300bbc0c3926354c6eeddbb953
Bug-Url: https://bugzilla.redhat.com/1351283
Signed-off-by: Sharon Gratch <sgratch@redhat.com>