webadmin: Hide hosted engine tab when editing host
The hosted engine tab in host editing or adding dialog
in the ui is only relevant when adding a new host.
This patch hides the tab when a host is edited.

To achieve this the initialization of the popup window
was moved after the host model initialization.
This allows setting the visibility for hostedEngineTab in
HostPopupView by using the getIsNew method (was already
implemented but didn't work since the relevant fields
weren't initialized).

Change-Id: I4746dda9fa0a06f0dec1d179639a20e5941e8ec9
Bug-Url: https://bugzilla.redhat.com/1350423
Signed-off-by: Jenny Tokar <jtokar@redhat.com>
(cherry picked from commit bb20845)