core: fix the serialization of vm statistics for gwt
VmStatistics class used to use collections that gwt cannot deserialize to
store the history of different statistics (cpu, memory, network). It was not
a problem before because the UI got VmStatistics that were retrieved from
the database and during the construction of the entities, collections that
can be deserialized by gwt were used instead. But now that were going to
send VmStatistics that are managed in-memory we must use collections that
can be deserialized by gwt.

Change-Id: I51b337e9d42a471bfd0cc11ad0fe9c2ca4ff9207
Bug-Url: https://bugzilla.redhat.com/1408725
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>