core: fix the serialization of vm statistics for gwt
VmStatistics class used to use collections that gwt cannot deserialize to
store the history of different statistics (cpu, memory, network). It was not
a problem before because the UI got VmStatistics that were retrieved from
the database and during the construction of the entities, collections that
can be deserialized by gwt were used instead. But now that were going to
send VmStatistics that are managed in-memory we must use collections that
can be deserialized by gwt.

Change-Id: Ibd0a60f24fda02a665edf5eb5c67e0766804b5c0
Signed-off-by: Arik Hadas <ahadas@redhat.com>