restapi: Don't try to modify @UriInfo parameters
Currently in the V3 compatibility layer we convert some body parameters
into query or matrix parameters, and to do so we modify the map of
values provided by the application server. But this map is unmodifiable
according to the documentation, and actually unmodifiable since version
3.0.17.Final of Resteasy, which is included in version 7.0.1 of JBoss
EAP. The net result is that when using that version of JBoss EAP all
attempts to convert body parameters fail. To avoid that this patch
introduces a new "parameters" field into the "Current" object. That will
be populated by the V3 layer, and consumed by the "ParametersHelper"
class that is reponsible for extracting parameters from different
places.

In order to make this change it is also necessary to move the
"ParametersHelper" and its callers to the "jaxrs" module, so that it can
use the "Current" class.

Change-Id: I3ebdc121772c3641230476ca9196dbabc20d1251
Bug-Url: https://bugzilla.redhat.com/1357296
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit ff63f5e4a22aeabd468a3fd443452cb56b40bc5b)