restapi: Add V3 creation server
In version 3 of the API there is a `creation_status` resource that
reports the status of asynchronously created resources. For example,
when creating a snapshot the result is looks like this:

  <snapshot id="456" href="/vms/123/snapshots/456">
    ...
    <link rel="creation_status" href="/vms/123/snapshots/456/creation_status/789"/>
    ...
  </snapshot>

The link points to a resource that can be used to check the creation of
the snapshot, for example:

  GET /vms/123/snapshots/456/creation_status/789

The result should be something like this:

  <creation id="789" href="vms/123/snapshots/456/creation_status/789">
    <status>
      <state>complete</state>
    </status>
  </creation>

The use of this is unreliable, as the operation may be synchronous or
asynchronous and return or not the status, depending on the
implementation details of the server.

The recommended way to check the status of the creation is to repeatedly
retrieve the snapshot and check its `snapshot_status` attribute.

However, we need to preserve the current behaviour in V3 of the API, and
currently this doesn't work, so we need to restore it.

Change-Id: Ife91560633294eae498522d19a189bfb11903028
Bug-Url: https://bugzilla.redhat.com/1359139
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>