aaa: Add support for dynamic porting
Add support to SSO for dynamic porting.
A new config variable has been added
ENGINE_SSO_INSTALLED_ON_ENGINE_HOST
that determines if SSO has been installed
on the same host as engine. By default
it is set to true and forwards the user
to the host and port that was used to access
engine.

Change-Id: I3d6aade0ca7126d93014ecda2e0f3abea4498395
Bug-Url: https://bugzilla.redhat.com/1362196
Signed-off-by: Ravi Nori <rnori@redhat.com>