restapi: Add V3 support for disks detail
Version 3 of the API supports a `detail` parameter that can be used to
include inline details into certain objects. In particular it can be
used to request details about the disks of a virtual machine:

  GET /ovirt-engine/api/vms
  Version: 3
  Accept: application/xml; detail=disks

This capability has been removed in version 4 of the API, but needs to
be preserved in version 3, and it works for all the objects that it used
to work, except for disks, because the "disks" collection has been
replaced with "diskattachments".

To fix this issue this patch moves the logic that handles the "detail"
parameter to the V3 compatibility layer, where the "disks" collection
still exists.

Change-Id: Icac0bea3503b542e60edf61fad5da0a1af9bac19
Bug-Url: https://bugzilla.redhat.com/1358860
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>