restapi: Improve affinity label resources
Currently the resources that manage the virtual machines and hosts that
are part of an affinity label are implemented using abstract generic
base classes. This doesn't work correctly with some versions of
Resteasy, as generic methods aren't recognized as implementations of the
methods annotated with @GET. In addition the methods to add virtual
machines and hosts aren't implemented correctly, as they return "action"
instead of the expected "vm" or "host" elements.

To address these issues this patch removes the abstract generic base
classes, and moves the reusable code to a new helper class.

The methods that add virtual machines and hosts are modified so that
they return "vm" or "host", as expected.

Change-Id: I274750970f283bb9ce813f301a28a5cbc544eb60
Bug-Url: https://bugzilla.redhat.com/1350228
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit d9981535b8e4ee9b7e0f4850805fd7aa5eaa65ee)