core: minor optimization in vm-analyzer
First check whether we get balloon information from VDSM and only
then check if balloon is enabled on the host. This reason for this
is that checking the host is a relatively costly operation in term
of CPU time (specifically because of the cloning of VDS object).

Not sure that the second check is needed anymore, but I prefer not
to remove it as part of this patch.

Change-Id: Ide71a7a5edcb44f05ca1b58e84d4ce3481fa5806
Signed-off-by: Arik Hadas <ahadas@redhat.com>