core: CreateAllSnapshotsFromVm - CreateAllCinderSnapshots execution
Currently CreateAllSnapshotsFromVm always executes CreateAllCinderSnapshots.
The execution may lead to FK constraint violation - this patch alters
the command to execute it only if the snapshot contains cinder disks.

This fix is for 3.6 branch only, as cinder support is on tech preview
on that version this patch fixes the issue for snapshot the includes image disks.
On the master branch the code is very different so this issue isn't relevant.

Change-Id: Iffb8fad4db3a2cea4000dff2491b40308cc2e986
Bug-Url: https://bugzilla.redhat.com/1357361
Signed-off-by: Liron Aravot <laravot@redhat.com>