webadmin:rename "SSL Verification" field
Rename "SSL Verification" checkbox field to
"verify server's SSL certificate" since it's more understandable
and also change the checkbox default value to checked.

Import from VMware provider uses the "SSL verification" check box to
determine if to add the "no_verify="1" flag to the URI or not:
1. if this checkox is checked (default) then the "no_verify=1" flag is not
added to the URI and a verification of the server's SSL certificate is
done.

2. If this checkbox is not checked then the "no_verify=1' flag is
added automatically to the URI and no VMware server's SSL certificate is
verified.

The current field is renamed to: "verify server's SSL
certificate".

The field renaming and default change is done for both
"Add/edit Provider" dialog (VMwarePropertiesWidget) and for Import
VMs dialog.

Change-Id: I4134da4f76e95354c0c980f31f4cb5a7b6d1d3cf
Bug-Url: https://bugzilla.redhat.com/1361193
Signed-off-by: Sharon Gratch <sgratch@redhat.com>