core: Fix match disk profile to storage domain
Problem cause:

A. When a template has a disk which is allocated to more than one
storage domain, it will contain for that disk image:
 - A list of storage domains
 - A list of disk profiles that belong to each storage domain

B. Creating a VM from that template using REST will :
 - Select the first storage domain from the list
 - Select the first disk profile from the list

Note : VM creation via UI selects a specific storage domain and a
specific disk profile which is associated with that storage domain.
There is no option to leave these values empty -
unlike a REST call where these values can be dismissed.

Consequence:
In cases where the selected disk profile is not associated with
the selected storage domain -
the VM will not be created due to this inconsistency.

Fix:
The disk profile is selected according to the selected
storage domain (out of the possible storage domains).

Change-Id: I2fd4b91842d8f8f6de5cb5e5a538e335c0fcfede
Bug-Url: https://bugzilla.redhat.com/1361838
Signed-off-by: yquinn <yquinn@redhat.com>