webadmin: userportal: Fix first vm creation in session
When creating a vm on a local storage domain it should be
pinned to a specific host that is available for this domain.

When creating a vm on local storage for the first time it wouldn't
be created with a pinned host due to an async code that would not run
in the expected order. When editing such a vm the user would get an
error since it seems that he is editing the pinned host.

To ensure the correct order of execution the handling of the local
storage was moved to the end of the block of code that is run
asynchronously.

Change-Id: I99918ef35099546734cf8835e5abe5bfcef33d08
Bug-Url: https://bugzilla.redhat.com/1284472
Signed-off-by: Jenny Tokar <jtokar@redhat.com>