packaging: drop FIX_RELEASE handling
When FIX_RELEASE was added to version.mak, the intention
was to allow to release an ovirt-engine package without
any modification to the java code, allowing to avoid to
bump the version in the pom files.

The intention was good but as per package maintenance we
keep aligned pom version and package version anyway when
we release. FIX_RELEASE has become meaningless and source
of human errors, therefore we're dropping it.

That done, POM_VERSION, APP_VERSION and RPM_VERSION
became redundant, dropped all of them and using the
unique value ENGINE_VERSION.
Kept only PACKAGE_VERSION which differ by the MILESTONE
suffix.

Change-Id: I4ff946551dc52d4a278976fa1cd049697c7496a3
Bug-Url: https://bugzilla.redhat.com/1354180
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>