webadmin: Web consoles warnings
Patternfly alert popup is in Spice HTML5 and noVNC if there is a problem
to connect to the websocket proxy server.

Warnings suggests to check that:
 * websocket proxy is running,
 * firewalls ares set
 * https is used (only for noVNC, Spice HTML5 can work with http)
 * app is accessed using predefined hostname (i.e. it doesn't work with
   e.g. numeric ip address as a hostname)
 * packages spice-html5/novnc are installed

Both Spice HTML5 and noVNC now works with http. Internally it uses
'wss'. Previously it tried to use 'ws' with 'http' however websocket
proxy server doesn't support 'ws'.

novnc-main.html and spicehtml5-main.html were changed to JSP to allow
includes to allow sharing html template. Encoding of *.jsp was set to
UTF-8 in 'services' module.

BrandingServlet was enabled for services module to get access to
Patternfly resources.

Change-Id: Ifdda870e13d77b24b95fe7109de377d560a0a86a
Bug-Url: https://bugzilla.redhat.com/1036221
Signed-off-by: Jakub Niedermertl <jniederm@redhat.com>