Extended Correlation ID capability and logging
This patch introduces the correlation id (CID) tracking for queries.
All incoming Rest API and Gwt RPC Api endpoints are now equipped with
the Correlation-Id header feature which allows an API user to assign a
CID in the request headers. If the header is not present, the engine
will generate for each request a new CID.
The used CID is passed back in the response as the Correlation-Id header
in the response.

Additionally to the header, the CID can now be also passed in as
correlation_id query parameter, which might be most interesting to the
REST Api

All queries are now logging the time spent and print the configured CID.

This feature has been added with the intend to allow better analysis of
slow tasks which come in through the REST API and/or the GWT RPC API.

Bug-Url: https://bugzilla.redhat.com/1347631
Change-Id: I67b45c14649bec21b2772f545bfde75aec9349b0
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>