webadmin: Don't call finazlize on image upload progress.
Currently, when a progress callback triggered while uploading an image,
and the upload status is not "TRANSFERRING", there is a call to finalize
upload. This is wrong, because in any scenario in which the status
becomes anything other than "TRANSFERRING", finalizing the upload
already happens. that means that currently a call for finalizing will
happen twice.

Remove the call for finalizeUpload in onProgress().

Change-Id: Iad2ad4b7500d44be2ff2ca9ea7ed91bc61c5c313
Bug-Url: https://bugzilla.redhat.com/1343878
Bug-Url: https://bugzilla.redhat.com/1328877
Signed-off-by: Amit Aviram <aaviram@redhat.com>