core+webadmin: Audit log a network error while uploading disk.
Currently when uploading a disk via webadmin, nothing warns the user
in case of a failure. e.g when the connectivity with
ovirt-imageio-proxy can't be established.

This scenario can happen if the proxy is down/not installed, or if
the user did not add the engine's CA as a certified authority in its
browser.

This patch adds AuditLog message to UploadImageStatusCommandParameters,
so if some error happens in the client side, it will append an
AuditLogMessge to the command, and this message will be logged and shown
to the user.

Also, handling the scenario of a network error in JS side is added in
this patch to demonstrate the new usage of the command and solve the
network error handling issue.

Change-Id: Ib7979ffec42dfa6b9f21ca74d4db7b393c2bff3c
Bug-Url: https://bugzilla.redhat.com/1343878
Bug-Url: https://bugzilla.redhat.com/1328877
Signed-off-by: Amit Aviram <aaviram@redhat.com>