core: AttachStorageDomainToPoolCommand - dont pass compensation context
If the data center is unintnalized, AttachStorageDomainToPoolCommand
executes the AddStoragePoolWithStorages command. Currently the
compensation context is passed to that command, on failure during the
command execution the data registered on the compensation context won't
be reverted becasue of an infrastractural bug.

Regardless of the need to fix the infrastractual issue in bz 1364804,
this patch  solves the bug by not passing the compensation context

Change-Id: I80c56009b005917bdb2b4a642fadef32623c59d0
Bug-Url: https://bugzilla.redhat.com/1359788
Signed-off-by: Liron Aravot <laravot@redhat.com>