database: memory_and_disk_images_storage_domain_view
Inline EXISTS conditions in a select list, should, in theory perform a
query once per row of the select list, while moving the same logic to a
LEFT JOIN should allow the database to perform the query just once and
hash-join it, resulting in a performance improvement.

This patch also adds the missing indexes on the snapshots table to make
these joins efficient.

According to the work done on verifying bug #1302752, on a loaded system
with 500 hosts,  7K VMs and 3 disks per VM, this patch improves the
query's runtime by a factor of 33 - from 6600 msecs to 200 msecs.

Change-Id: Ifbedd49da10fa6752234eacf72bbb0f350b3c778
Related-To: https://bugzilla.redhat.com/1363823
Related-To: https://bugzilla.redhat.com/1302752
Signed-off-by: Allon Mureinik <amureini@redhat.com>