engine: Fix taskhandler init - RemoveDisksSnapshot
The method initTaskHandlers is called every time that the command is
initiated.

In RemoveDiskSnapshotsCommand, the number of task handlers created was
different after that the former handler finished its operation.
This situation created a mismatch between the task handler map in
BaseCommand and the command execution index.

In this patch, the images IDs are computed only once and not every time
that the initTaskHandlers is invoked, keeping the number of task
handlers consistent.

Also, the task handlers are not created if it is a live merge flow.

Change-Id: Ic2836d7f6430183eecef337ed6d1c667b2f2990c
Bug-Url: https://bugzilla.redhat.com/1365393
Signed-off-by: Fred Rolland <frolland@redhat.com>