restapi: Ignore failures fetching cluster policy
Currently when a cluster is fetched in version 3 of the API an
additional query is performed to retrieve the details of its associated
scheduling policy, as that is how it used to work. But this additional
query may require additional permissions for some users, and when this
happens the complete cluster fetch fails. To alleviate that problem this
patch changes the API so that failures to fetch the scheduling policy
will just be ignored. As most users won't be interersted in those
details this should solve most use cases. When a user without the
required permissions fetches the cluster she will get only the
reference, and not the details:

  GET /clusters
  Authorization: myuser@mydomain:...
  Filter: true

  <cluster id="123" href="/clusters/123">
    ...
    <scheduling_policy id="456" href="/schedulingpolicies/456"/>
    ...
  </cluster>

When the user has the required permissions he will get all the details
of the scheduling policy.

Change-Id: I23b0ef3830f8291d725fa8fce241c8e0c12309be
Bug-Url: https://bugzilla.redhat.com/1364048
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>