webadmin: v2v- kvm username and password are not mandatory
The user, password fields are no longer mandatory.
This is supported by adding a checkbox named "authenticaton requires"
as exists in the external providers dialog.

The user name validation rules are also changed by ommiting the check
for special chars and domain.

Furthermore, add a tooltip with a short explanation on the KVM URI field.

Change-Id: I807559e74d18dd37f746b5f4eface1850f4cd6f0
Bug-Url: https://bugzilla.redhat.com/1351542
Signed-off-by: Sharon Gratch <sgratch@redhat.com>