webadmin: Make Reports-related Dashboard main tab initially hidden
In 4.0, even though we don't ship Reports anymore, UI still contains
code for Reports integration: (Reports related) "Dashboard" main tab
will be shown on the right-hand side of tab panel for a brief moment,
until the UI detects unavailability of Reports & hiding that tab.

This is unrelated to the "Dashboard" main tab shown on the left-hand
side of tab panel, representing the oVirt Dashboard feature introduced
in 4.0.

This patch simply sets the (Reports related) "Dashboard" main tab to
be initially hidden, which fixes the above mentioned UX issue. Due to
Reports not available in 4.0, that tab will never become visible.

Change-Id: I155ac581eb71febb55df84b339a6ab46b729dde7
Bug-Url: https://bugzilla.redhat.com/1365980
Signed-off-by: Vojtech Szocs <vszocs@redhat.com>