webadmin: re-enlarge the create snapshot message dialog
In case there are no disks allowing a snapshot for the VM, the message
panel size should be enlarged to contain the following two messages:
"Could not detect Guest Agent on the VM. Note that without a Guest Agent
the data on the created snapshot may be inconsistent.
There are no disks allowing a snapshot, only the configuration will be a
part of the VM snapshot.".

And therefore the whole dialog should be enlarged also.

Change-Id: Ib28fad0bbaea9dc16163a9f0ecd71f35668762a6
Bug-Url: https://bugzilla.redhat.com/1360914
Signed-off-by: Sharon Gratch <sgratch@redhat.com>