restapi: Fix V3 for force remove data center
In version 3 of the API it was possible to remove a data center
specifying the "force" parameter in the request body:

  DELETE /datacenters/123
  Content-Type: application/xml

  <action>
    <force>true</true>
  </action>

But in version 4 this was replaced with a parameter, because DELETE
requests wity bodies are forbidden:

  DELETE /datacenters/123?force=true

But we need to preserve the old behaviour for backwards compatibility.
That is what this patch does.

Change-Id: Ibf04e0f19f109c287319f4e4f6bcc6e9d5218f86
Bug-Url: https://bugzilla.redhat.com/1366794
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>