restapi: Update capabilities for V3
The "capabilities" provided by version 4 of the engine for version 3 of
the API aren't generated directly, instead they are stored in a
"capabilities.xml" file that is extracted from version 3 of the engine.
This file needs to be updated when there are relevant changes in version
3 of the engine. This patch updates it, using version 3.6.7.5 of the
engine.

Change-Id: Ic3085a972be693c5638005f0acaeab58233bb077
Bug-Url: https://bugzilla.redhat.com/1367400
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>
(cherry picked from commit 69939e4e8b071653c6ba3232cb133b4edf0351fe)