core: AsyncTasks - use persisted context instead of generating one
When all the async tasks of command ends, CommandAsyncTask executes
the command endAction(). It uses ExecutionHandler.createFinalizingContext()
to create the finalizing context for the command.

For 'traditional' async task manager flows, this behavior worked fine - that's
because the parent always executed the endAction() of the child commands by itself.
On COCO flows the endAction() of a child commands might be executed when it ends by
the ATM - that causes for the FINALIZING step to be added on that phase to the job
instead of on the parent end.

Currently we cache the command contexts, so there's no need to get the context
from createFinalizingContext(), we can use the exact context the command was executed
with and by that to avoid the problem.

Change-Id: I065551ee471a314ffd84b0f72a7b6079156e7f7a
Signed-off-by: Liron Aravot <laravot@redhat.com>