core: StorageJobCallback - execute endAction()
This patch adds execution of the endAction() method from the StorageJobCallback
to perform the standard endAction() operations for StorageJob commands as well.
Untill now, that wasn't relevant for those commands, as now we intend them
to add command steps (added in commit 6833ee94) we need to execute the endAction
to end those steps.

Change-Id: I5dc940acd60cd956ce1905902b84773636b67120
Signed-off-by: Liron Aravot <laravot@redhat.com>