aaa: Request state does not match session state after successful login
Fix the usability issues with mismatched
session state when the login screen session
has expired.

The client id and client secret with the
redirect uri check should be sufficient
security check, session "state" which is
a random string and was passed between sso
and engine as an additional level of security
can be dropped with out any security threats.

Change-Id: I9874c007e2d3382bbcdc8a280302306e2e6dc601
Bug-Url: https://bugzilla.redhat.com/1367921
Signed-off-by: Ravi Nori <rnori@redhat.com>