restapi: Update to model 4.0.32
The more relevant changes included in the new version of the model are
the following:

* Add the `clone_permissions` parameter to the operations that add
  virtual machines and templates.

* Add the `clone` parameter to the operation that adds a new virtual
  machine.

* Add the `ad_partner_mac` attribute to the bonding type.

* Add the `ad_aggregator_id` attribute to the host NIC type.

* Add the `ExternalVmImportsService` and related types.

* Add the `active` property back to the disk type.

Note that all those changes were actually introduced in version 4.0.31
of the model, but there was a small bug affecting the SDKs but not the
engine. Fixing that bug required making a new 4.0.32 release, and it is
better to keep the engine in sync with the latest model.

Change-Id: Iaf4ec501f1065ebb8a949c6cc3eb9f0a1315aee5
Related-To: https://bugzilla.redhat.com/1281666
Related-To: https://bugzilla.redhat.com/1346847
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>