restapi: Don't duplicate Maven exec plugin
Currently we use the Maven exec plugin several times, to generate the
RSDL and also to generate the Java code from the model of the API. To do
so we specify the plugin twice in the POM files. Newer versions of Maven
don't accept that, and the build fails. To avoid that problem this patch
replaces the two plugin usages with one single usage with multiple
executions.

Change-Id: I81a4d8d48cc5dc15e18048f10fae5c82ae4d4d18
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>