engine: Add audit log message for update OVF_STORE
Since OVF_STORE update can now be initiated by the user, it's only
natural an event will be logged on a successfull completion.
This patch adds an audit log message for the user-initiated update use
case. The quartz-scheduled update is not audit logged, in order to void
flooding the events log with messages.

Change-Id: I2aed2c85f5896a22a6037d831f311cbb1302b4ff
Bug-Url: https://bugzilla.redhat.com/1270562
Signed-off-by: Vered Volansky <vvolansk@redhat.com>
Signed-off-by: Tal Nisan <tnisan@redhat.com>