core: Adding maximum limitation for image chain.
Due to the maximum path length, snapshots operations start to
fail around the 98th image in the chain.

Currently, this is an expected failure as the actual link to
the underlying image grows as it walks the chain.

When that point is reached, snapshots fail and the VM won't
migrate/restart.

The engine should limit that number of images for a chain to avoid
failures

Change-Id: Ib8607dd84bc789c3478aa8d008bedeae40e21a95
Bug-Url: https://bugzilla.redhat.com/1348405
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>