added support for isRunOnlyIfAllValidation flag
Added support for running the runMultipleAction() method with the
isRunOnlyIfAllValidation flag.

The isRunOnlyIfAllValidation flag existed in code but was not supported
(i.e. was always treated as set to false)

It was used by Gluster with flag set to true and therefore changed to
false to avoid regression and should be
tested by them after this fix.

Change-Id: Ic3ae6bfd5929d0fd804dae4f55d18e9dd0b1eb8e
Bug-Url: https://bugzilla.redhat.com/1364171
Signed-off-by: Sharon Gratch <sgratch@redhat.com>