webadmin: v2v- avoid closing wizard if vm's validation failed
In case that one or more VMs failed in validation phase (due to
non unique name in data center or any other validation failure),
a warning dialog is displayed with relevant message and when clicking
"ok", the v2v wizard is displayed back in its last state and not closed
as it was before.

Notes:
1. The VM name uniquness is per Data Center and therefore the relevant
message was rephrased to inform that the VM already exists in Data Center
and not in the whole system.
2. in case that at least one of the VMs failed in validation phase, all
valid VMs import will also stopped from proceeding.
In such a case, a message is displayed and the user can go back to v2v
wizard to fix it. The import will start for all VMs only in case all are
valid.

Change-Id: I82bf607127322caf4ef246c25cd9b1848192ce18
Bug-Url: https://bugzilla.redhat.com/1364171
Signed-off-by: Sharon Gratch <sgratch@redhat.com>