core: update lun's info on any property change
When syncing DB LUNs with the information from the storage, if the LUNs'
IDs are different we save a new LUN, and if the LUN's size was changed,
we update the existing one.

This approach ignores other fields that may have changed, so that:
* If another field which is not the LUN's size was changed, the LUN will
  not be updated.
* If the LUN's size was changed, the lun will be updated, along with
  other fields that were changed.

This patch takes a new approach - getLunsToUpdateInDb returns the LUNs
that require a DB update and updateLunsInDb always updates these LUNs
in the DB (or saves new ones if the ID was changed).
Therefore, adding more filters to getLunsToUpdateInDb (like LUNId and
deviceSize) will possibly cause more LUNs to be updated in the DB.

Change-Id: I9cfd28001fe065001669040ba010d2ee54ac876d
Signed-off-by: Idan Shaby <ishaby@redhat.com>