restapi: Ignore failures fetching additional disk data
When disks are retrieved using version 3 of the API additional data is
required, like the disk attachments. If the user doesn't have permission for
the additional data then the complete operation fails. This is a problem for
users that may not have permission for that additional data. In order to avoid
that this patch changes the API so that failures to retrieve the additional
data are just ignored.

Change-Id: I37159aaa3d767df6c83e666bfcd70ad4ff503e87
Bug-Url: https://bugzilla.redhat.com/1368552
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>