restapi: Don't add V3 guest IP to links
Currently the V3 compatibility layer tries to add the guest IP to all
the returned objects of type "vm". That is wrong, as links, for example,
shouldn't include such information. To solve the problem this patch
changes the API so that the population of the guest IP is performed only
for objects of type "vm" returned by the "list", "get", and "update"
operations of the services that manage virtual machines.

Change-Id: I68bfbb75cb3ef733625b2b11a9f744a664a69cda
Related: https://bugzilla.redhat.com/1368552
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>