restapi: Don't add V3 disk attachment details to links
Currently the V3 compatibility layer tries to add the attachment details
(bootable, etc) to all the returned objects of type "disk". That is
wrong, as links, for example, shoundn't include such information. To
solve the problem this patch changes the API so that the population of
these details is performed only for objects of type "disk" returned by
the "list", "get", and "update" operations of the services that manage
virtual machine disks.

Change-Id: Ic8bc7e4b498d3d7ee7d09b60432b08f12be669c3
Related: https://bugzilla.redhat.com/1368552
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>