core: adding RemoveDeviceFromSANStorageDomainCommand
This patch adds the RemoveDeviceFromSANStorageDomainCommand, the command
will monitor and track the flow of removing the device from a storage
domain.
At first, the command will execute the MoveStorageDomainDevice command
to move the data from the removed device to other devices of the storage
domain. After the data was moved from the device the command will
execute the ReduceStorageDomainDevice command to reduce the device from
the domain.

Change-Id: I3d68ffc1a373fb328c3229362f6888150b6e2eb2
Bug-Url: http://bugzilla.redhat.com/1403839
Signed-off-by: Liron Aravot <laravot@redhat.com>