core: log reason for stopping vm only if set
Previously we always tried to include the stop reason when auditing
VM stop operation. When no reason was set, either because it is not
enabled in the cluster or the user simply did not enter a reason,
the audit log included the following redundant message:
(Reason: No reason was returned for this operation failure.
 See logs for further details.)

This patch changes that so audit logs will include the stop reason
only when it is set. That way, we avoid messages as shown above.

Change-Id: I25b3ffced6036796d20bfcce0d47c22a3365a80d
Bug-Url: https://bugzilla.redhat.com/1340414
Signed-off-by: Arik Hadas <ahadas@redhat.com>