packaging: spec: require newest otopi
Require otopi >= 1.6.0 which is targeted to 4.1.
It includes a new feature which allow to get coverage
reports when executing the setup.

In order to enable the feature you'll need to provide
a configuration file like:

8<--------------------------------------
[run]
branch = True

[report]
exclude_lines =
    # Have to re-enable the standard pragma
    pragma: no cover

    # Don't complain about missing debug-only code:
    def __repr__
    if self\.debug

    # Don't complain if tests don't hit defensive assertion code:
    raise AssertionError
    raise NotImplementedError

    # Don't complain if non-runnable code isn't run:
    if 0:
    if __name__ == .__main__.:

ignore_errors = True
8<------------------------------------

Be sure to have python-coverage and libselinux-python
installed on your system and then execute:

  export OTOPI_COVERAGE=1
  export COVERAGE_PROCESS_START="automation/coverage.rc"
  engine-setup

In order to generate the report:
  mkdir coverage_html_report
  coverage html -d coverage_html_report

Change-Id: I41df99d82723e998eda9084b93063cd842d8f065
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>