core: Set hosted engine maintenance mode to none after upgrade
When upgrading hosted engine host, the engine first moves it
to maintenance mode in the engine and sets local maintenance
to true on the host itself.
After finishing the upgrade the engine sets the host status to
up but doesn't change the local maintenance status of the host.
This fix sets the local maintenance to false after upgrade if the
host is a hosted engine host.

Change-Id: I48f193b99dc8e71aef7e36532d36c551f9927ecd
Bug-Url: https://bugzilla.redhat.com/1370907
Signed-off-by: Jenny Tokar <jtokar@redhat.com>
(cherry picked from commit aa05b05)