backend: add missing commands' Guid constructors
The constructor that is used as a part of the compensation process on
startup is the one that receives the command id (Guid), and it should
exist in each class that is annotated with the
'NonTransactiveCommandAttribute' annotation and the annotation's
'forceCompensation' attribute is set to true.

This patch adds these missing constructors and a test to enforce their
existence in the future.

Change-Id: Ic369352b45ce86b3d38b9feb18d02d9d2ed80e9f
Signed-off-by: Idan Shaby <ishaby@redhat.com>