image upload: don't expose resource ID to entity updates.
Resource ID is used for identifying the upload tickets. once the
ability of clearing it is exposed for the frontend/REST, it can
potentially cause loss of a transfer session, opening up security risks.

Remove this member from the image transfer updates entity, and set it as
a parameter for the backend ImageTransferUpdater class.

Change-Id: Iee39b43faea82a4737919de0c39acba4b2b60b26
Signed-off-by: Amit Aviram <aaviram@redhat.com>