backend: call getVgInfo after extending a storage domain
When extending a storage domain, we get the new luns' information from
vdsm when calling ConnectAllHostsToLun (in validate()) and save it to
the db. This causes an inconsistency between the luns' real info from
the storage and their info as we save it in the db, for example a
different physical volume id.

Calling getVgInfo right after the extend itself gets us the new info,
and saving the new info fixes this bug.

Change-Id: I6a987f1f81f6ecad56a6854090fbba21131cbabd
Bug-Url: https://bugzilla.redhat.com/1371960
Signed-off-by: Idan Shaby <ishaby@redhat.com>