core: HttpImageTaskVDSCommand - problematic type inference
In change Iaed648bcf979d682389a4df92, the explicit Integer type was
removed from the call to Config.getValue(), that causes the type to
be infered as Long and fails on ClassCastException.

When the type is explictly specified, the returned value is an Integer
which is outboxed and can be passed to future.get().

Change-Id: I349a08a7d119bcc57f8f6a88331c9ea64790cc99
Bug-Url: https://bugzilla.redhat.com/1371833
Signed-off-by: Liron Aravot <laravot@redhat.com>