webadmin: Mark instance image as changed in new and attach
Change I998e8c573c8afaf45668729cf15fb0ebe3edbad5 introduced a change where
disk update for instance images will be invoked only upon edit, however,
this made new and attach disk actions to be skipped as well.
This patch adds the indication of a change to create and attach disk as
well so the will not be skipped.

Change-Id: I093ccc18a4ab70b4f2a5bcd3ef46abeb41cee55c
Bug-Url: https://bugzilla.redhat.com/1371164
Signed-off-by: Tal Nisan <tnisan@redhat.com>