core: adding first_metadata_device to storage domain static
We added a support to VDSM to report the device of the first extant
of the domain metadata lv.
VDSM relies on the fact that the metadata lv first extant is always 0 -
therefore we should block operations for being done on that device.
For example, attempt to move the data from that device might change
the first extant of the metadata lv to something different than 0 -
making the domaun unusable.

While VDSM code has validation to prevent such operations from being
performed, we should have that information on the engine side to provide
better user experience by blocking attempts to send such requests.

NOTE: The metadata LV might be extended to reside on more devices,
the column saves the device of the first extant of the lv.

Change-Id: I54e35f7cb3b51725c219f625a3d3434575d66113
Signed-off-by: Liron Aravot <laravot@redhat.com>